+++ wazo_calld/exceptions.py
@@ -136,17 +136,3 @@ class WazoConfdUnreachable(APIException):
                 'original_error': str(error),
             }
         )
-
-
-class InvalidExtension(APIException):
-
-    def __init__(self, context, exten):
-        super().__init__(
-            status_code=400,
-            message='Invalid extension',
-            error_id='invalid-extension',
-            details={
-                'context': context,
-                'exten': exten,
-            }
-        )
