+++ wazo_calld/plugins/applications/services.py
@@ -8,7 +8,7 @@ from requests import HTTPError
 from ari.exceptions import ARINotFound
 from wazo_calld.plugin_helpers import ami, confd
 from wazo_calld.plugin_helpers.ari_ import Channel
-from wazo_calld.exceptions import InvalidExtension
+from wazo_calld.plugin_helpers.exceptions import InvalidExtension
 from .models import (
     CallFormatter,
     make_node_from_bridge,
