+++ wazo_calld/plugins/transfers/services.py
@@ -7,11 +7,11 @@ from ari.exceptions import ARINotFound
 from xivo.caller_id import assemble_caller_id
 
 from wazo_calld.ari_ import DEFAULT_APPLICATION_NAME
-from wazo_calld.exceptions import InvalidExtension
 from wazo_calld.plugin_helpers import ami
 from wazo_calld.plugin_helpers.ari_ import Channel
 from wazo_calld.plugin_helpers.confd import User
 from wazo_calld.plugin_helpers.exceptions import (
+    InvalidExtension,
     NotEnoughChannels,
     TooManyChannels,
     UserPermissionDenied,
